start of CLI
This is not ready to be merged yet, just looking for initial feedback. This is the start of the CLI using argparse. Plans are to include the options to choose between load_confounds or niak_load_confounds, easier bash interface, options for model saving and output file, and improvement of file list implementation. 

Feedback and suggestions welcome! Please do not merge this as is!